Bug 1198458: Rollup of changes previously applied to media/webrtc/trunk/webrtc and fixes to those rs=jesup r=froyd,jib,bwc,jesup,gcp,sotaro,pkerr,pehrsons Landing as one rolled-up patch to avoid breaking regression tests, and in keeping with previous WebRTC imports.  Broken out parts that needed review are on the bug.

+++ media/webrtc/signaling/src/media-conduit/WebrtcMediaCodecVP8VideoCodec.h
@@ -45,7 +45,7 @@ public:
 
   virtual int32_t Release() override;
 
-  virtual int32_t SetChannelParameters(uint32_t packetLoss, int rtt) override;
+  virtual int32_t SetChannelParameters(uint32_t packetLoss, int64_t rtt) override;
 
   virtual int32_t SetRates(uint32_t newBitRate, uint32_t frameRate) override;
